Check if the model went well

check_model computes tests to assess if the model went well. It is important to run this step before going ahead with the analysis otherwise you may make mistakes in the interpretation of the results.

check_model(x)

Arguments

x
outputs from

model_GxE

model_bh_intra_location

model_bh_GxE

model_spatial

model_bh_variance_intra

model_hedonic

model_napping

Details

S3 method. See for more details :

check_model.fit_model_GxE regarding model_GxE

check_model.fit_model_bh_intra_location regarding model_bh_intra_location

check_model.fit_model_bh_GxE regarding model_bh_GxE

check_model.fit_model_spatial regarding model_spatial

check_model.fit_model_bh_variance_intra regarding model_bh_variance_intra

check_model.fit_model_hedonic regarding model_hedonic

check_model.fit_model_napping regarding model_napping

See also