Simulated data set for unipart network for seed lots which represents relation between between seed lots

A dataset containing four locations, 15 germplasms and 5 years. Relation between seed lots are diffusion, mixture, reproduction and selection. The data set contain the following variables: seed_lot_parent, seed_lot_child, relation_type, relation_year_start, relation_year_end, germplasm_parent, location_parent, year_parent, alt_parent, long_parent, lat_parent, germplasm_child, location_child, year_child, alt_child, long_child, lat_child

See the vignette for more information.

data(data_network_unipart_sl)

Format

A data frame with 104 rows and 17 columns