Check and format the data to be used by PPBstats functions

format_data_PPBstats checks and formats the data to be used by PPBstats functions

format_data_PPBstats(data, type, threshold, network_part = c("unipart",
  "bipart"), network_split = c("germplasm", "relation_year_start"),
  vertex_type = NULL)

Arguments

data
The data frame to format

type
type of format :

data_network

data_agro

data_agro_SR

data_agro_HA

data_agro_LF

data_organo_napping

data_organo_hedonic

threshold
For type = data_organo, number of occurence of descriptors <= threshold are kept

network_part
For type = "data_network", element of the network, it can be "unipart" or "bipart"

network_split
For type = "data_network" and network_part = "unipart" and vertex_type = "location", split of the data that can be "germplasm" or "relation_year_start"

vertex_type
For type = "data_network",

for unipart network : "seed_lots" or "location"

for bipart network : c("germplasm", "location")

Details

See for more details :

format_data_PPBstats.data_network

format_data_PPBstats.data_agro

format_data_PPBstats.data_agro_SR

format_data_PPBstats.data_agro_HA

format_data_PPBstats.data_agro_LF

format_data_PPBstats.data_organo_napping

format_data_PPBstats.data_organo_hedonic