Check and format the data to be used by PPBstats functions for hedonic analyses

format_data_PPBstats.data_organo_hedonic checks and formats the data to be used by PPBstats functions for hedonic analyses

format_data_PPBstats.data_organo_hedonic(data, threshold)

Arguments

data
The data frame with the following columns : sample, juges, note, descriptors, germplasm, location. The descriptors must be separated by ";". Any other column can be added as supplementary variables.

threshold
number of occurence of descriptors <= threshold are kept

Value

It returns a list of four elements

data the data formated to run the anova and the multivariate analysis regarding

sample

sample means

juges

var_sup the supplementary variables used in the multivariate analysis

descriptors the vector of descriptors cited knowing the threshold applyed when formated the data.

Details

See the book for more details here.

See also