Get the value of type one error needed to have X groups.

get.at.least.X.groups gets the value of type one error needed to have X groups. X goes from 2 to the maximum number of groups possible.

get.at.least.X.groups(Mpvalue, MCMC, p.adj = "soft.bonf",
  precision = 5e-04)

Arguments

Mpvalue
A vector or a matrix with the probabilities for two parameters to be equal. It comes from the function comp.parameters

MCMC
MCMC. It is a data frame.

p.adj
NULL for no adjustement of the type one error. "soft.bonf" for a soft bonferonni correction to take into account multiple comparisons (alpha / number of parameters).

precision
The precision of the type one error with the corresponding groups. The smaller the better, but the smaller the more time consuming due to computing matters

Value

The function returns a vector with the value of type one error for each group

Details

This function is helpful to get at least two (or more) groups with a given type one error.

When there is no data, it returns NA. It may depend on the precision. For exemple, if you go from 5 groups to 2 groups, it will be NA for 4 groups and 3 groups because the precision was not small enought.

See also