Get regional farms data and satellite farms data

get.env.info look at a data set and separate regional farms and satellite farms in two datasets. This function is used in model_bh_intra_location.

get.env.info(D, nb_ind)

Arguments

D
The data frame with its specific column names used in model_bh_intra_location

nb_ind
The minimal number of individuals per variable

Value

The function returns a list with

"vec_env_with_no_data": a vector with the environments without data for the given variable

"vec_env_with_no_controls": a vector with the environments with no controls

"vec_env_with_controls": a vector with the environments with controls

"vec_env_RF": a vector with the environments as regional farms (i.e. with at least two blocks)

"vec_env_SF": a vector with the environments as satellite farms (i.e. with one block)

"D_RF": a data frame with regional farms

"D_SF": a data frame with satellite farms

See also