Return a "mean vs stability" ggplot based on PCA object

ggplot_mean_vs_stability returns a "mean vs stability" ggplot based on PCA object

ggplot_mean_vs_stability(res.pca)

Arguments

res.pca
output from biplot_data.check_model_GxE

Value

A ggplot object.

Details

It is a list of two elements where the plots are drawn under the same construction. A red circle indicates the position of the average location, which is defined by the average Dim1 and Dim2 scores across all locations. An arrow points this average 'virtual' location. A red line passes through the biplot origin and the average location (known as the Average-Tester Axis (ATA)). A red line is set perpendicular the ATA. Dashed perpendicular lines to the ATA are drawn and passed through each germplasm.

mean_performance : assess mean performance score of each germplasm. The mean performance score is the rank of germplasm along the ATA. The arrow points the greatest value according to their mean performance across all locations. An high score, represents an high mean performance. A low score reresents an low mean performance. Score are displayed into the legend.

stability_performance : assess stability performance score of each germplasm. The stability performance score of the germplasms are represented by the projection from the germplasm to the ATA. An high score, represents an high interaction and therfore lower stability. A low score reresents an low interaction and therefore an high stability. Score are displayed into the legend.

See also