Run Hierarchical Bayesian GxE model

model_bh_GxE runs Hierarchical Bayesian GxE modelto get main germplasm, environment and sensitivity effects over the network

model_bh_GxE(data, variable, nb_iterations = 1e+05, thin = 10,
  return.alpha = TRUE, return.sigma_alpha = TRUE, return.beta = TRUE,
  return.sigma_beta = TRUE, return.theta = TRUE,
  return.sigma_theta = TRUE, return.epsilon = FALSE,
  return.sigma_epsilon = TRUE, return.DIC = FALSE)

Arguments

data
The data frame on which the model is run. It should come from format_data_PPBstats.data_agro

variable
The variable on which runs the model

nb_iterations
Number of iterations of the MCMC

thin
thinning interval to reduce autocorrelations between samples of the MCMC

return.alpha
Return the value for each germplasm main effect (alpha_i)

return.sigma_alpha
Return the value of the variance of the distribution where the alpha_i come from

return.beta
Return the value for each sensitivity to environments (beta_i)

return.sigma_beta
Return the value of the variance of the distribution where the beta_i come from

return.theta
Return the value for each environment main effect (theta_j)

return.sigma_theta
Return the value of the variance of the distribution where the theta_j come from

return.epsilon
Return the value of the residuals of the model (epsilon_ij)

return.sigma_epsilon
Return the value of the variance of the distribution where the epsilon_ij come from

return.DIC
Return the DIC value of the model. See details for more informations.

Value

The function returns a list with

"data.presence.absence.matrix": a matrix germplasm x environment with the number of occurence in the data

"model.presence.absence.matrix": a matrix germplasm x environment with the number of occurence in the data used for the model (i.e. with at least two germplasm by environments.)

"germplasm.not.used": the vector of germplasms not used in the analysis because they were not on at least two environments. If NULL, all the germplasms were used in the analysis.

"MCMC": a list with the two MCMC chains (mcmc object) from the model

"epsilon": a vector with the median value of the epsilon_ij

"DIC": the DIC value of the model

Details

Hierarchical Bayesian GxE model estimates germplasm (alpha_i), environment (theta_j) and sensitivity to interaction (beta_i) effects. An environment is a combinaison of a location and a year.

The different effects are taken in different distributions of respective variances sigma_alpha, sigma_theta and sigma_beta. This model takes into acount all the information on the network in order to cope with the high disequilibrium in the dataset (i.e. high percentage of missing GxE combinaisons on the network).

First, the additive model is done. This model gives intitial values of some parameters of the Hierarchical Finlay Wilkinson model which is done next.

The model is run on data set where germplasms are on at least two environments.

More information can be found in the book: https://priviere.github.io/PPBstats_book/family-2.html#model-2

For DIC value, see ?dic.samples from the rjags package for more information.

References

P. Riviere, J.C. Dawson, I. Goldringer, and O. David. Hierarchical multiplicative modeling of genotype x environment interaction for flexible experiments in decentralized participatory plant breeding. In prep, 2015.

See also